chore: refactor vcs icon
