+++ frontend/src/views/SettingWorkspaceVCS.vue
@@ -32,9 +32,9 @@
 <script lang="ts">
 import { reactive, computed, watchEffect, defineComponent } from "vue";
 import { useRouter } from "vue-router";
+import VCSCard from "@/components/VCS/VCSCard.vue";
+import VCSSetupWizard from "@/components/VCS/VCSSetupWizard.vue";
 import { useVCSV1Store } from "@/store";
-import VCSCard from "../components/VCSCard.vue";
-import VCSSetupWizard from "../components/VCSSetupWizard.vue";
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface LocalState {}
